First commit of change_q330_annc
This script will review the ANNC structure on a q330 and
as a default, will update the POC settings with the settings
in the change_q330_annc.pf

Use with caution!  Read the man page!  Do not run it without
the -n option until you are certain of what you are doing!!!