Enable fast_lookup to improve performance.
This is a fairly major change including an interface change for
GCLgrid3d and all of it's children.   When adapting pwmig to
not use antelope experiments showed the cautious method used
in the old lookup method was very expensive in computer time.
pwmig run times were reduced by a factor for 2 to 4 by disabling
the recover feature in that code.   The fast method that does not
use that recovery method is now the default.   The cautious method
is still available as an option through two new methods that enable
or disable the fast lookup method.