Parameterize various hardcodes in q330_baler
The netcode was available in the parameter file before, but it wasn't
being used.

The Institution was hard-coded in the email subject line.

The netcode and Institution are available in site.pf typically, so use
a &ref call in the default parameter file

The default parameter file value of status_email had production email
addresses in it.