Update for 5.5, clean up internals
Updated the dbwfserver to work with Antelope 5.5. Handle changes to
paths as a result of the contrib relocation

Remove global imports of everything from __main__ into submodules,
instead relying on each submodule to do it's own imports. This required
handling the config object as a parameter instead of a global variable
inside of the resource class.

Renamed several Class names to have a Capital First Letter to cut down
on confustion between instances variables and class names.