Prune stachan pairs that aren't in wfdisc
since we're already looking at every single row in the wfdisc, we might
as well use that to prune the list of stachans so that we don't include
channels without data. This should eliminate the need for some of the
custom stachan extraction routines that were created to work around this
problem.