Upgrade dbmoment code to newer version of the tdmt_invc_iso code.
This update will also upgrade the methods used for the creation of
the synthetics needed, taking them from 8 elements to 10 elements.
This will require new synthetic databases for dbmoment users.
The requirement for previous magnitudes was eliminated by introducing
a dynamic search for the best filter that will reduce the variance
of the signal. Any new number of filters could be specify in the PF
file. Multiple models could also be specify in the PF file and only
the best solution will be kept.
A new method of jackknife recursive search was implemented to help
identify the best set of stations for the inversion.
New logging capabilities added with the capacity to save the logs
to a file in a dedicated directory.
New inversion code used in this version:
Douglas S. Dreger and Sean Ford June 3, 2011
Time-Domain Moment Tensor INVerse Code (tdmt_invc_iso)
Release 3.0